+++ src/content-scripts/components/ViewOnOpenSaucedButton/ViewOnOpenSaucedButton.ts
@@ -4,6 +4,7 @@ import { createHtmlElement } from "../../../utils/createHtmlElement";
 
 export const ViewOnOpenSaucedButton = (username: string) => {
   const viewOnOpenSaucedButton = createHtmlElement("a", {
+    id: "view-on-opensauced-button",
     href: `https://insights.opensauced.pizza/user/${username}/contributions`,
     className:
       "inline-block mt-4 text-black bg-gh-white dark:bg-gh-gray dark:text-white rounded-md p-2 text-sm font-semibold text-center select-none w-full border hover:shadow-button hover:no-underline",