+++ src/pages/start.tsx
@@ -2,7 +2,7 @@ import OpenSaucedLogo from "../assets/opensauced-logo.svg";
 import { SUPABASE_LOGIN_URL } from "../constants";
 
 const Start = () => (
-    <div>
+    <div className="p-4 bg-slate-800">
       <img
         alt="Open Sauced Logo"
         src={OpenSaucedLogo}
@@ -30,7 +30,7 @@ const Start = () => (
         className="bg-orange no-underline border-none rounded-md text-white font-bold py-2 px-4 cursor-pointer
           bg-gradient-to-r from-[#e67e22] to-[#d35400]"
       >
-        Login!
+        Login
       </a>
     </div>
   );